+++ pages/app/datasets/[datasetId]/versions/[versionName]/index.tsx
@@ -1,7 +1,7 @@
 import { NewContext } from "../../../../../../lib/appLocalContext";
 import { getDatasetBy } from "../../../../../../lib/dataset";
 import { getUserByUID } from "../../../../../../lib/users";
-import DatasetDetailsPage from "../../../DatasetDetailsPage";
+import DatasetDetailsPage from "../../../../../../components/DatasetDetailsPage";
 
 export default function DatasetDetailsByVersionNamePage(props) {
   return DatasetDetailsPage(props);
